+++ shelfsense/src/components/products/ProductsList.jsx
@@ -1,7 +1,7 @@
 import {Skeleton, Stack} from '@mui/material';
 import Grid from '@mui/material/Grid2';
 import { useEffect, useState } from 'react';
-import { fetchProducts } from '../../services/productService';
+import { fetchProducts } from '../../util/services/productService';
 import ProductsCard from "./ProductsCard";
 
 const ProductsList = () => {
@@ -39,7 +39,7 @@ const ProductsList = () => {
             <>
                 {loading ? (
                     Array.from({ length: 10 }).map((_, index) => (
-                        <Grid item xs={12} md={6} key={index}>
+                        <Grid xs={12} md={6} key={index.toString()}>
                             <Stack spacing={1}>
                                 <Skeleton animation="wave" variant="rectangular" width={210} height={120} />
                                 <Skeleton variant="text" width={120} />
@@ -49,7 +49,7 @@ const ProductsList = () => {
                     ))
                 ) : (
                     products.map((product) => (
-                        <Grid item xs={12} sm={6} md={4} lg={3} key={product.id} sx={{mg: '10'}}>
+                        <Grid xs={12} sm={6} md={4} lg={3} key={product.id} sx={{mg: '10'}}>
                             <ProductsCard product={product} />
                         </Grid>
                     ))
