+++ shelfsense/src/pages/LoginPage.jsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { Box, Button, Card, CardContent, TextField, Typography, CircularProgress } from '@mui/material';
-import { login } from '../services/userService.jsx';
+import { login } from '../util/services/userService.jsx';
 import { useNavigate } from 'react-router-dom';
 
 const LoginPage = () => {
