+++ shelfsense/src/util/services/componentService.jsx
@@ -3,8 +3,12 @@ const COMPONENTS_URL = `${import.meta.env.VITE_API_URL}/components`;
 
 export const fetchComponents = async () => {
     try {
-        const response = await fetch(COMPONENTS_URL);
-        if (!response.ok) throw new Error('Failed to fetch components');
+        const response = await fetch(COMPONENTS_URL, {
+            credentials: "include",
+        });
+        if (!response.ok) {
+            throw new Error('Failed to fetch components');
+        }
         console.log(response);
         return await response.json();
     } catch (error) {
