+++ shelfsense/src/util/services/productService.jsx
@@ -3,8 +3,12 @@ const PRODUCTS_URL = `${import.meta.env.VITE_API_URL}/products`;
 
 export const fetchProducts = async () => {
     try {
-        const response = await fetch(PRODUCTS_URL);
-        if (!response.ok) throw new Error('Failed to fetch products');
+        const response = await fetch(PRODUCTS_URL, {
+            credentials: "include",
+        });
+        if (!response.ok) {
+            throw new Error('Failed to fetch products');
+        }
         console.log(response);
         return await response.json();
     } catch (error) {
